+++ engine/server/engine/centralized_logs/logs_database_client.go
@@ -6,11 +6,16 @@ import (
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 )
 
+type ConjunctiveLogLineFilters interface {
+	GetConjunctiveLogLineFiltersString() string
+}
+
 type LogsDatabaseClient interface {
 	GetUserServiceLogs(
 		ctx context.Context,
 		enclaveID enclave.EnclaveID,
 		userServiceGuids map[service.ServiceGUID]bool,
+		conjunctiveLogLineFilters ConjunctiveLogLineFilters,
 	) (
 		userServiceLogsByServiceGuidChan chan map[service.ServiceGUID][]LogLine,
 		errChan chan error,
@@ -21,6 +26,7 @@ type LogsDatabaseClient interface {
 		ctx context.Context,
 		enclaveID enclave.EnclaveID,
 		userServiceGuids map[service.ServiceGUID]bool,
+		conjunctiveLogLineFilters ConjunctiveLogLineFilters,
 	) (
 		userServiceLogsByServiceGuidChan chan map[service.ServiceGUID][]LogLine,
 		errChan chan error,